+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - World stats contain separate reserved/used memory fields (#177)
 - Provides `MapX` for up to 12 components (#182)
 - Adds `MapX.Set` and `Map.Set` (#183)
+- Adds function `ResourceTypeID` (#184)
 
 ### Documentation